+++ HISTORY.rst
@@ -14,6 +14,7 @@ Unreleased
 **Features and Improvements**
 
 * ALCN-1055: add last suite name on the partner
+* ALCN-1054: Add tag OnlinePayment and change rule StatisticCode on customer xml export
 
 **Bugfixes**